requirements.txt
# ─── Основные ───────────────────────────────
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
fpdf2>=2.7.0

# ─── Фоллбэк PDF (опционально) ─────────────
reportlab>=3.6.0

# ─── Проверка библиотеки (опционально) ──────
# PyMuPDF>=1.23.0

# ─── Сборка EXE (опционально) ───────────────
# pyinstaller>=6.0.0